opt_client_serv/lab7/requirements.txt
bcrypt==5.0.0
grpcio==1.76.0
grpcio-tools==1.76.0
python-jose==3.5.0
